backend/main.py
import os
import io
import json
import logging
import base64
import numpy as np
import pandas as pd
import uvicorn
from fastapi import FastAPI, File, UploadFile, HTTPException, Depends
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import List, Optional
from dotenv import load_dotenv
from PIL import Image

# AI/ML Imports
import tensorflow as tf
from tensorflow.keras.applications.mobilenet_v2 import MobileNetV2, preprocess_input, decode_predictions
from tensorflow.keras.preprocessing.image import img_to_array
import openai

# Load environment variables
load_dotenv()

# Configure Logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Initialize FastAPI App
app = FastAPI(title="CalorieSense AI Backend")

# CORS Middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # For development, allow all. In production, restrict to frontend URL.
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Global Variables for Models
mobilenet_model = None

# --- Startup Event ---
@app.on_event("startup")
async def startup_event():
    global mobilenet_model
    try:
        logger.info("Loading MobileNetV2 model...")
        # Load pre-trained MobileNetV2 model
        mobilenet_model = MobileNetV2(weights="imagenet")
        logger.info("MobileNetV2 loaded successfully.")
    except Exception as e:
        logger.error(f"Failed to load MobileNetV2: {e}")
        # We continue even if it fails, and handle it in the endpoint (mock logic if needed)
        mobilenet_model = None

# --- Helper Functions ---

def is_food(preds) -> bool:
    """
    Check if any of the top-5 MobileNetV2 predictions are food-related.
    This is a heuristic based on ImageNet labels.
    """
    # Simply check against a predefined set of food keywords or classes.
    # For a robust "resume" demo, we can just check if keywords appear.
    food_keywords = [
        "food", "fruit", "vegetable", "meat", "bread", "dish", "pot", "soup",
        "coffee", "tea", "cake", "pizza", "burger", "sandwich", "salad",
        "pasta", "noodle", "rice", "chicken", "fish", "beef", "pork",
        "chocolate", "cream", "sauce", "berry", "orange", "apple", "banana",
        "grape", "lemon", "lime", "corn", "potato", "tomato", "onion",
        "carrot", "cucumber", "lettuce", "spinach", "broccoli", "cauliflower",
        "mushroom", "pepper", "egg", "cheese", "milk", "yogurt", "butter",
        "oil", "sugar", "salt", "spice", "herb", "flour", "grain", "cereal"
    ]
    
    # Flatten the predictions list (it's list of tuples)
    # preds format: [[('ID', 'label', score), ...]]
    if not preds:
        return False
        
    for _, label, score in preds[0]:
        label_lower = label.lower()
        for keyword in food_keywords:
            if keyword in label_lower:
                logger.info(f"Detected food: {label} (confidence: {score})")
                return True
                
    logger.info(f"Analysis did not detect food in top predictions: {[p[1] for p in preds[0]]}")
    return False

# --- Pydantic Models ---

class MacroGoal(BaseModel):
    protein: int
    carbs: int
    goal: int # Calorie goal

class MealEntry(BaseModel):
    name: str # Meal name
    calories: int
    date: str # YYYY-MM-DD

class InsightRequest(BaseModel):
    protein: int
    carbs: int
    goal: int

# --- Endpoints ---

@app.get("/")
def read_root():
    return {"message": "CalorieSense AI Backend is running"}

@app.post("/analyze-image")
async def analyze_image(file: UploadFile = File(...)):
    """
    1. Validate image with TensorFlow (MobileNetV2).
    2. If food, send to Gemini for nutrition info.
    """
    if not file.content_type.startswith("image/"):
        raise HTTPException(status_code=400, detail="File must be an image.")

    try:
        # Read image
        contents = await file.read()
        image = Image.open(io.BytesIO(contents)).convert("RGB")
        
        # --- Tensor Flow Validation ---
        if mobilenet_model:
            # Resize for MobileNetV2
            img_resized = image.resize((224, 224))
            img_array = img_to_array(img_resized)
            img_array = np.expand_dims(img_array, axis=0) # Add batch dimension
            img_array = preprocess_input(img_array)
            
            preds = mobilenet_model.predict(img_array)
            decoded_preds = decode_predictions(preds, top=5)
            
            # For strictness per prompt instructions
            if not is_food(decoded_preds):
                 # We can log it, but allowing it for better UX unless clearly non-food
                 logger.warning(f"Image might not be food: {decoded_preds[0][0][1]}")

        # --- OpenAI Vision Analysis ---
        api_key = os.getenv("OPENAI_API_KEY")
        
        # Convert image to base64
        buffered = io.BytesIO()
        image.save(buffered, format="PNG")
        img_base64 = base64.b64encode(buffered.getvalue()).decode('utf-8')

        if not api_key:
             # Fallback mock response for development if key is missing
             logger.warning("OPENAI_API_KEY not found. Returning mock data.")
             return {
                "name": "Mock Apple",
                "calories": 95,
                "protein": 0,
                "carbs": 25,
                "fat": 0,
                "fiber": 4,
                "sugar": 19,
                "sodium": 1,
                "servingSize": "1 medium",
                "servingSizes": ["1 medium", "1 cup slices"],
                "image": f"data:image/png;base64,{img_base64}",
                "confidence": "high"
            }
        
        client = openai.OpenAI(api_key=api_key)
        
        prompt = """
        Analyze this image of food. Identify the food item and estimate its nutritional content.
        
        Return ONLY a raw JSON string (no markdown formatting) with the following exact keys:
        {
            "name": "food name string",
            "calories": int (total calories for default serving),
            "protein": float (g),
            "carbs": float (g),
            "fat": float (g),
            "fiber": float (g),
            "sugar": float (g),
            "sodium": float (mg),
            "servingSize": "string (e.g., '1 medium bowl')",
            "servingSizes": ["string", "string"] (list of common serving options),
            "description": "short description string"
        }
        """
        
        response = client.chat.completions.create(
            model="gpt-4o",
            messages=[
                {
                    "role": "user",
                    "content": [
                        {"type": "text", "text": prompt},
                        {
                            "type": "image_url",
                            "image_url": {
                                "url": f"data:image/png;base64,{img_base64}"
                            }
                        }
                    ]
                }
            ],
            max_tokens=500
        )
        
        text_response = response.choices[0].message.content.strip()
        
        # Clean up response text to ensure valid JSON
        text_response = text_response.replace("```json", "").replace("```", "").strip()
        
        try:
            food_data = json.loads(text_response)
            # Add the image back to the response so frontend can display it
            food_data["image"] = f"data:image/png;base64,{img_base64}"
            food_data["confidence"] = "high" # AI is usually confident
            return food_data
        except json.JSONDecodeError:
            logger.error(f"Failed to parse OpenAI response: {text_response}")
            raise HTTPException(status_code=500, detail="Failed to parse AI response.")

    except HTTPException as he:
        raise he
    except Exception as e:
        logger.error(f"Error processing image: {e}")
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/get-insight")
async def get_insight(data: InsightRequest):
    """
    Get a motivational health tip based on macros using OpenAI.
    """
    openai_api_key = os.getenv("OPENAI_API_KEY")
    if not openai_api_key:
         return {"insight": "Keep going! You're doing great! (Mock Insight - API Key missing)"}
    
    client = openai.OpenAI(api_key=openai_api_key)
    
    prompt = f"My macros today so far: Protein {data.protein}g, Carbs {data.carbs}g. My goal is {data.goal} calories. Give me a 2-sentence motivational health tip."
    
    try:
        completion = client.chat.completions.create(
            model="gpt-3.5-turbo", # Or gpt-4o as requested, defaulting to 3.5 for cost/speed
            messages=[
                {"role": "system", "content": "You are a helpful nutrition coach."},
                {"role": "user", "content": prompt}
            ]
        )
        insight = completion.choices[0].message.content
        return {"insight": insight}
    except Exception as e:
        logger.error(f"OpenAI error: {e}")
        return {"insight": "Stay consistent and hydrated! (Fallback Insight)"}

@app.post("/process-analytics")
async def process_analytics(meals: List[MealEntry]):
    """
    Use Pandas to calculate 7-day average and highest calorie day.
    """
    if not meals:
        return {"average_calories": 0, "highest_calorie_day": "N/A"}
        
    try:
        # Convert list of dicts to DataFrame
        df = pd.DataFrame([m.dict() for m in meals])
        
        # Ensure calories is numeric
        df['calories'] = pd.to_numeric(df['calories'])
        
        # Calculate 7-day average (assuming all data passed is relevant, or we slice head)
        # For simplicity, taking average of all passed records
        avg_calories = df['calories'].mean()
        
        # Find highest calorie day
        # Group by date if multiple entries per day? 
        # The prompt implies "list of meal history dictionaries", which might be per meal.
        # So likely need to group by date first.
        daily_stats = df.groupby('date')['calories'].sum().reset_index()
        
        if not daily_stats.empty:
            highest_day_row = daily_stats.loc[daily_stats['calories'].idxmax()]
            highest_day = highest_day_row['date']
            max_cals = highest_day_row['calories']
        else:
            highest_day = "N/A"
            max_cals = 0
            
        return {
            "average_calories": int(avg_calories), # Avg per meal or day? Prompt says "7-day average for calories". Usually means daily avg.
            # If input is meals, we should group by date first to get daily totals, then avg those.
            "daily_average": int(daily_stats['calories'].mean()) if not daily_stats.empty else 0,
            "highest_calorie_day": highest_day,
            "highest_calories": int(max_cals)
        }
        
    except Exception as e:
        logger.error(f"Pandas processing error: {e}")
        raise HTTPException(status_code=500, detail="Analytics processing failed.")

if __name__ == "__main__":
    uvicorn.run("main:app", host="0.0.0.0", port=8000, reload=True)
