backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
tensorflow
pandas
openai
python-dotenv
pillow
numpy
requests
